==> 02_CorpusCollection.ipynb <==
## Scraping MSFT News on Twitter

from twitterscraper import query_tweets
import datetime as dt
import pandas as pd
import os

begin_date=dt.date(2019,1,1)
end_date=dt.date(2019,12,31)
limit=100000
lang='english'

tweets=query_tweets("#MSFT OR #$MSFT OR $MSFT OR MSFT OR Microsoft from:wsj OR from:reuters OR from:business OR from:cnbc OR from:RANsquawk OR from:wsjmarkets",
                    limit=limit,lang=lang)

StockNewsTweets =pd.DataFrame(t.__dict__ for t in tweets)
df = StockNewsTweets.copy()

df = df[['timestamp', 'text', 'username']]
df.set_index(pd.to_datetime(df['timestamp']), inplace=True)
df.drop(columns='timestamp', inplace=True)

df

df.drop_duplicates(subset=['text'], keep=False, inplace=True)
df

# Check if stock news data exit
if not os.path.exists('./msft_news_twitter.csv'):

# make directory and store stock news    
    #os.makedirs('stockdata')
    df.to_csv('./msft_news_twitter.csv')
else:
    pass

trOnly_tweets=query_tweets("#MSFT OR #$MSFT OR $MSFT OR MSFT OR Microsoft from:reuters",lang=lang)

StockNewsTweets_tr =pd.DataFrame(t.__dict__ for t in tr_tweets)
df = StockNewsTweets_tr.copy()
df

StockNewsTweets_trOnly =pd.DataFrame(t.__dict__ for t in trOnly_tweets)
df = StockNewsTweets_trOnly.copy()
df

twitterscraper "#MSFT OR #$MSFT OR $MSFT OR MSFT OR Microsoft from:wsj OR from:reuters OR from:business OR from:cnbc OR from:RANsquawk OR from:wsjmarkets" -o msft_tweets.json -l 1000000 --lang en -p 500 -ow

df_orig = pd.read_json('msft_tweets.json', encoding='utf-8')
df = df_orig.copy()

df = df[['timestamp', 'text', 'username']]
df
