==> 03-Data Preprocessing.ipynb <==
#### The script below is to put together different news source, delete duplicates, and slice news between 2015 and 2019

import datetime as dt
import pandas as pd
import numpy as np
import os
import re

df_orig1 = pd.read_json('msft_tweets.json', encoding='utf-8')
a_df = df_orig1.copy()
a_msft_df = pd.DataFrame(a_df[['timestamp', 'text']])
a_msft_df

df_orig2 = pd.read_json('msft_tweets_old.json', encoding='utf-8')
b_df = df_orig2.copy()
b_msft_df = pd.DataFrame(b_df[['timestamp', 'text']])
b_msft_df

df_orig3 = pd.read_json('msft_tweets_old2.json', encoding='utf-8')
c_df = df_orig3.copy()
c_msft_df = pd.DataFrame(c_df[['timestamp', 'text']])
c_msft_df

# Merge twitter news
msft_tweeets = pd.concat([a_msft_df, b_msft_df, c_msft_df],ignore_index=True)

msft_tweeets

msft_tweeets['Date'] = msft_tweeets['timestamp'].dt.date

msft_tt = msft_tweeets.copy()

msft_tt.rename(columns={"text": "News"}, inplace = True)
msft_tt.drop(columns=['timestamp'], inplace=True)

msft_tt.sort_values(by=['Date'], inplace=True)

msft_tt.reset_index(drop=True)

news_a = pd.read_csv('msft_news.csv')
newsline_a = news_a.copy()
newsline_a['Date1'] = newsline_a['Date'].apply(lambda x: dt.datetime.strptime(x, '%B %d, %Y')).dt.date
msft_nla = pd.DataFrame(newsline_a[['Date1', 'News']])

msft_nla.rename(columns={"Date1": "Date"}, inplace = True)
msft_nla.sort_values(by=['Date'], inplace=True)
msft_nla.reset_index(drop=True)

news_b = pd.read_csv('msft_news_twitter.csv')
newsline_b = news_b.copy()
newsline_b['Date'] = newsline_b['timestamp'].apply(lambda x: dt.datetime.strptime(x, '%Y-%m-%d %H:%M:%S')).dt.date
msft_nlb = pd.DataFrame(newsline_b[['Date', 'text']])

msft_nlb.rename(columns={"text": "News"}, inplace = True)
msft_nlb.sort_values(by=['Date'], inplace=True)
msft_nlb.reset_index(drop=True)

news_c = pd.read_csv('msft_news2.csv')
newsline_c = news_c.copy()
newsline_c['Date'] = newsline_c['timestamp'].apply(lambda x: dt.datetime.strptime(x, '%Y-%m-%d %H:%M')).dt.date
msft_nlc = pd.DataFrame(newsline_c[['Date', 'text']])

msft_nlc.rename(columns={"text": "News"}, inplace = True)
msft_nlc.sort_values(by=['Date'], inplace=True)
msft_nlc.reset_index(drop=True)

# Merge reuter's news
msft_reuter = pd.concat([msft_nla, msft_nlb, msft_nlc],ignore_index=True)

msft_reuter.sort_values(by=['Date'], inplace=True)

msft_reuter

msft_tt = msft_tt[['Date', 'News']]

# Merge all news, use date as index, and delete duplicates
msft_allnews = pd.concat([msft_reuter, msft_tt],ignore_index=True)
msft_allnews.set_index(['Date'], inplace=True)
msft_allnews.drop_duplicates(subset=['News'], keep=False, inplace=True)

msft_allnews.sort_values(by=['Date'], inplace=True)
msft_allnews

# Extract master list of news headlines and save as csv
startdate = pd.to_datetime("2015-01-01").date()
enddate = pd.to_datetime("2019-12-31").date()

msft_masterNews = msft_allnews.loc[startdate:enddate]
msft_masterNews

# Check if stock news data exit
if not os.path.exists('./msft_masternews.csv'):

# make directory and store stock news    
    #os.makedirs('stockdata')
    msft_masterNews.to_csv('./msft_masternews.csv')
else:
    pass

msft_nhl = pd.read_csv('msft_headlines05.csv')
msft_nhl

msft_nhl['News1'] = msft_nhl['News'].str.replace('...', '', regex=False)

msft_nhl.drop(['News'], axis=1, inplace=True)

msft_nhl.rename(columns={"News1": "News"}, inplace = True)

#msft_nhl.set_index(['Date'], inplace=True)
msft_nhl.drop_duplicates(subset=['News'], keep=False, inplace=True)

msft_nhl

# Check if stock news data exit
if not os.path.exists('./msft_headlines.csv'):

# make directory and store stock news    
    #os.makedirs('stockdata')
    msft_nhl.to_csv('./msft_headlines.csv')
else:
    pass
