==> 08_TransE.ipynb <==
### TransE model implemented using Ampligraph library

import ampligraph
ampligraph.__version__

import numpy as np
import pandas as pd
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt

import tensorflow as tf
from ampligraph.latent_features import TransE
from ampligraph.evaluation import evaluate_performance
from ampligraph.latent_features import save_model, restore_model
from ampligraph.evaluation import mr_score, mrr_score, hits_at_n_score
from ampligraph.datasets import load_from_csv
from ampligraph.evaluation import train_test_split_no_unseen 

# load data
tec_kg = load_from_csv("data\KGE","tec.txt",sep=', ')

all_tec = pd.read_csv('./data/KGE/tec_entity.txt',sep=", ", header=None, engine='python')
Entity = all_tec.values

X_train, X_test = train_test_split_no_unseen(tec_kg, test_size=0.20)

X_train

# Initialise TransE model
model = TransE(batches_count=100, 
               epochs=50, 
               k=20, 
               loss='multiclass_nll',
               eta=2,
               optimizer_params={'lr':1e-3},
               optimizer='adam',
               verbose=True,
              )

positives_filter = tec_kg 

# Fitting the model
tf.logging.set_verbosity(tf.logging.ERROR)

model.fit(X_train, early_stopping = False)

# Calibrate the output probabilities
# model.calibrate(kgdata['valid'], positive_base_rate=0.5)

# Saving model 
save_model(model, 'TransE_model.pkl')

# Load model
transe_model = restore_model('TransE_model.pkl')

# Evaluate model
ranks = evaluate_performance(X_test, 
                             model=transe_model, 
                             filter_triples= positives_filter,   # Corruption strategy filter defined above 
                             use_default_protocol=True, # corrupt subj and obj separately while evaluating
                             verbose=True)

# Metrics
mrr = mrr_score(ranks)
print("MRR: %.2f" % (mrr))

hits_10 = hits_at_n_score(ranks, n=10)
print("Hits@10: %.2f" % (hits_10))
hits_3 = hits_at_n_score(ranks, n=3)
print("Hits@3: %.2f" % (hits_3))
hits_1 = hits_at_n_score(ranks, n=1)
print("Hits@1: %.2f" % (hits_1))

Entities = transe_model.get_embeddings(Entity)

Entities.shape

# Project embeddings into 2D space via PCA
head_pca = PCA(n_components=2)
tail_pca = PCA(n_components=2)
entity_pca1 = head_pca.fit_transform(Entities[:,0])
entity_pca2 = tail_pca.fit_transform(Entities[:,1])

head_pca.explained_variance_ratio_, tail_pca.explained_variance_ratio_

entity_pca1

entity_pca2

te_embed = pd.DataFrame(entity_pca1, columns = ['ent1PC1', 'ent1PC2'])
te2_embed = pd.DataFrame(entity_pca2, columns = ['ent2PC1', 'ent2PC2'])

te_embd = pd.concat([te_embed,te2_embed], axis=1)

te_embd.to_csv('te_embedding.csv')
