==> 09_TCN.ipynb <==
### TCN Stock Prediction Model developed

import os
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report, confusion_matrix 
from sklearn.metrics import roc_curve, roc_auc_score, f1_score, accuracy_score

import tensorflow.keras.backend as K
from tensorflow.keras import Input, Model
from tensorflow.keras.layers import Conv1D,Conv2D, Dense, Dropout
from tensorflow.keras.layers import MaxPool2D
from tcn import TCN, tcn_full_summary, compiled_tcn

#### TCN model on KGML Dataset

# Load KGML Data
kg_data = pd.read_csv('kg_complete.csv')

kg_data.head()

# Split dataset in features (X) and labels (y)
x = kg_data.drop(columns = ['Target'])
x = np.array(x)
y = np.ravel(kg_data['Target'])

f = x.reshape(x.shape[0], x.shape[1], 1)
l = y

f.shape, l.shape

# Split KG into training and test sets
x_train,x_test,y_train,y_test = train_test_split(f, l, test_size=0.2, 
                                                 random_state=42)

x_train.shape, x_test.shape, y_train.shape, y_test.shape

# Initialise model

i = Input(shape=(6,1), name='Input')
x = Conv1D(128,6,name='Conv1D_layer')(i)
x = TCN(nb_filters=64,
        kernel_size=6,
        dilations=[1, 2, 4, 8],
        nb_stacks=1,
        padding='causal',
        dropout_rate=0.2,
        activation='relu',
        #lr=0.002,
        name='TCN_layer')(x)
x = Dense(32, activation='relu', name='Dense_layer' )(x)
x = Dropout(0.2)(x)
x = Dense(1, activation='sigmoid', name='output')(x)

model = Model(inputs=[i], outputs=[x])
model.compile('adam', 'binary_crossentropy', metrics=['accuracy', 'mse'])
model.summary()

model.fit(x_train, y_train, epochs = 100, verbose=1)

?model.save

pred = model.predict(x_test)
y_pred = np.where(pred <=0.5, 0, 1)

# Metrics
pred_accuracy = accuracy_score(y_test, y_pred)
fscore = f1_score(y_test, y_pred)
roc_auc = roc_auc_score(y_test, y_pred)
print('accuracy_score: %.3f' % pred_accuracy)
print('f1_score: %.3f' % fscore)
#print('roc_auc: %.3f' % roc_auc)

### TCN Model on TransE Dataset

# Load TransE Data
te_complete = pd.read_csv('te_complete.csv')

te_complete.head()

# Split dataset in features (X) and labels (y)
xt = te_complete.drop(columns = ['Target'])
xt = np.array(xt)
yt = np.ravel(te_complete['Target'])

m = xt.reshape(xt.shape[0], xt.shape[1], 1)
n = yt

m.shape, n.shape

# Split KG into training and test sets
xt_train, xt_test, yt_train, yt_test = train_test_split(m, n, 
                                                        test_size=0.2,
                                                        random_state=42)

xt_train.shape, xt_test.shape, yt_train.shape, yt_test.shape

# Initialise model

it = Input(shape=(8,1), name='Input')
xt = Conv1D(128,8,name='Conv1D_layer')(it)
xt = TCN(nb_filters=64,
        kernel_size=6,
        dilations=[1, 2, 4, 8],
        nb_stacks=1,
        padding='causal',
        dropout_rate=0.2,
        activation='relu',
        #lr=0.002,
        name='TCN_layer')(xt)
xt = Dense(32, activation='relu', name='Dense_layer' )(xt)
xt = Dropout(0.2)(xt)
xt = Dense(1, activation='sigmoid', name='output')(xt)

transE_model = Model(inputs=[it], outputs=[xt])
transE_model.compile('adam', 'binary_crossentropy', 
                     metrics=['accuracy', 'mse'])
transE_model.summary()

transE_model.fit(xt_train, yt_train, epochs = 100, verbose=1)

tpred = transE_model.predict(xt_test)
yt_pred = np.where(tpred <=0.5, 0, 1)

# Metrics
pred_accuracy = accuracy_score(yt_test, yt_pred)
fscore = f1_score(yt_test, yt_pred)
roc_auc = roc_auc_score(yt_test, yt_pred)
print('accuracy_score: %.3f' % pred_accuracy)
print('f1_score: %.3f' % fscore)
#print('roc_auc: %.3f' % roc_auc)

### TCN Model on Baseline Dataset

# Load baseline data
bl_complete = pd.read_csv('bl_merged.csv')

bl_complete.head()

# Split dataset in features (X) and labels (y)
xb = bl_complete.drop(columns = ['Target'])
xb = np.array(xb)
yb = np.ravel(te_complete['Target'])

u = xb.reshape(xb.shape[0], xb.shape[1], 1)
v = yb

u.shape, v.shape

# Split KG into training and test sets
xb_train, xb_test, yb_train, yb_test = train_test_split(u, v, 
                                                        test_size=0.2,
                                                        random_state=42)

xb_train.shape, xb_test.shape, yb_train.shape, yb_test.shape

# Initialise baseline model

ib = Input(shape=(4,1), name='Input')
xb = Conv1D(128,4,name='Conv1D_layer')(ib)
xb = TCN(nb_filters=64,
        kernel_size=6,
        dilations=[1, 2, 4, 8],
        nb_stacks=1,
        padding='causal',
        dropout_rate=0.2,
        activation='relu',
        #lr=0.002,
        name='TCN_layer')(xb)
xb = Dense(32, activation='relu', name='Dense_layer' )(xb)
xb = Dropout(0.2)(xb)
xb = Dense(1, activation='sigmoid', name='output')(xb)

bl_model = Model(inputs=[ib], outputs=[xb])
bl_model.compile('adam', 'binary_crossentropy', 
                     metrics=['accuracy', 'mse'])
bl_model.summary()

bl_model.fit(xb_train, yb_train, epochs = 100, verbose=1)

bpred = bl_model.predict(xb_test)
yb_pred = np.where(bpred <=0.5, 0, 1)

# Metrics
bpred_accuracy = accuracy_score(yb_test, yb_pred)
bfscore = f1_score(yb_test, yb_pred)
broc_auc = roc_auc_score(yb_test, yb_pred)
print('accuracy_score: %.3f' % bpred_accuracy)
print('f1_score: %.3f' % bfscore)
#print('roc_auc: %.3f' % broc_auc)
